+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ### Added
-- Support for Racket
+- Support for Racket (Thanks to @Ghexor)
+- Support for F# (Thanks to @chlohal)
 
 
 ## [1.5.0]